refactor: move types in personal-details/hooks.ts to ~/types.ts
